:recycle: Move duplicated scope handling to method
This method could also be utilised in all the Begin/Process/End paths.
However, they each have a very slightly different version of the code,
do things in different orders, and split up certain parts of the code.

We can revisit whether we want to refactor those code paths to use the
new method in future, but for now this avoids duplicating the code from
here to DlrScriptCommandProcessor, which would introduce complications.